Move AbstractCrawlServiceServlet to gwtp-crawler module.
