+++ gwtp-core/gwtp-crawler/pom.xml
@@ -36,6 +36,11 @@
             <artifactId>servlet-api</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>net.sourceforge.htmlunit</groupId>
+            <artifactId>htmlunit</artifactId>
+        </dependency>
+
         <!-- DI dependencies -->
         <dependency>
             <groupId>com.google.inject</groupId>
