+++ gwtp-core/pom.xml
@@ -251,6 +251,13 @@
                 <version>${velocity.version}</version>
             </dependency>
 
+            <!-- HTML Dependencies -->
+            <dependency>
+                <groupId>net.sourceforge.htmlunit</groupId>
+                <artifactId>htmlunit</artifactId>
+                <version>${htmlunit.version}</version>
+            </dependency>
+
             <!-- Test dependencies -->
             <dependency>
                 <groupId>org.jukito</groupId>
